+++ js/data.js
@@ -1,0 +1,28 @@
+const description = [
+  'Кто сдесь?',
+  'Как-то так...',
+  'Многабукав мало толку...',
+  'Оупс...',
+  'Кагдила?',
+  'Вот эт я понимаю...',
+];
+
+const message = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const authorName = [
+  'Вася',
+  'Мессер',
+  'Максим',
+  'Тёщя',
+  'Лейла',
+  'Лана',
+];
+
+export {description, message, authorName};
